backend/requirements.txt
Flask
numpy
psycopg2-binary
python-dotenv
flask-sqlalchemy
cockroachdb
Flask-Login
sqlalchemy-cockroachdb==2.0.1
